Handle Voted PO2 from CCR Dive Computers
This commit attempts to integrate the dive computer's logged
displayed/voted PO2 so that it can be logged and/or analysed.

Some dive computers, Shearwater units in particular, store individual
sensor values as uncalibrated and don't provide calibration data. These
computers do, however, store the PO2 reported to the diver during the
course of the dive, based on voting logic.

This commit carries that "voted" PO2 through into the data models
and saved files so that it can be stored and graphed. The CCR sensor
averaging function in Subsurface has been modified so that if Voted PO2
is available (greater than 0), the Voted PO2 value will be used as the
PO2 in the dive profile graph. If it isn't present Subsurface will
calculate the PO2 based on existing code.

The Voted PO2 is carried across as a separate value into logbook saves.

This commit relates to 1edcdf8b7ffdfd3834714e1b9b4408711f9613cd in
libdivecomputer.

Signed-off-by: Aaron Scheiner <blue@aquarat.co.za>